refactor: fixes import of account cubit
